41650: moved files and made small full item page changes
