+++ src/app/item-page/item-page-routing.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { ItemPageComponent } from './item-page.component';
+import { ItemPageComponent } from './simple/item-page.component';
+import { FullItemPageComponent } from './full/full-item-page.component';
 
 @NgModule({
     imports: [
         RouterModule.forChild([
             { path: 'items/:id', pathMatch: 'full', component: ItemPageComponent },
+            { path: 'items/:id/full', pathMatch: 'full', component: FullItemPageComponent },
         ])
     ]
 })
