--- src/app/item-page/item-page.component.scss
@@ -1,1 +1,0 @@
-@import '../../styles/variables.scss';
