+++ src/app/item-page/simple/item-page.component.scss
@@ -1,0 +1,1 @@
+@import '../../../styles/variables';
