+++ src/app/item-page/simple/metadata-field-wrapper/metadata-field-wrapper.component.ts
