+++ src/app/item-page/simple/metadata-uri-values/metadata-uri-values.component.ts
