+++ src/app/item-page/simple/metadata-values/metadata-values.component.ts
